Move shared-module semver resolution out of generated modules map
